Add macro for generating ProtocolConfig
