+++ protocols/mp-ecdsa/src/lib.rs
@@ -1,8 +1,7 @@
 use gadget_common::client::{AccountId, ClientWithApi, PalletSubmitter};
-use gadget_common::config::ProtocolConfig;
 use gadget_common::debug_logger::DebugLogger;
+use gadget_common::define_protocol;
 use gadget_common::gadget::network::Network;
-use gadget_common::gadget::WebbGadgetProtocol;
 use gadget_common::keystore::{ECDSAKeyStore, KeystoreBackend};
 use pallet_jobs_rpc_runtime_api::JobsApi;
 use sc_client_api::Backend;
@@ -52,12 +51,7 @@ where
 
     logger.info("Done creating keygen protocol");
 
-    let protocol = EcdsaProtocolConfig {
-        network: Some(network),
-        client: Some(client_inner),
-        protocol: Some(protocol),
-        _pd: std::marker::PhantomData,
-    };
+    let protocol = EcdsaProtocolConfig::new(network, client_inner, protocol);
 
     gadget_common::run_protocol(protocol).await?;
     Ok(())
@@ -95,12 +89,7 @@ where
 
     logger.info("Done creating sign protocol");
 
-    let protocol = EcdsaProtocolConfig {
-        network: Some(network),
-        client: Some(client_inner),
-        protocol: Some(protocol),
-        _pd: std::marker::PhantomData,
-    };
+    let protocol = EcdsaProtocolConfig::new(network, client_inner, protocol);
 
     gadget_common::run_protocol(protocol).await?;
     Ok(())
@@ -152,38 +141,4 @@ where
     }
 }
 
-pub struct EcdsaProtocolConfig<N, P, C, B: Block, BE> {
-    pub network: Option<N>,
-    pub client: Option<C>,
-    pub protocol: Option<P>,
-    _pd: std::marker::PhantomData<(B, BE)>,
-}
-
-impl<
-        B: Block,
-        BE: Backend<B> + 'static,
-        N: Network,
-        C: ClientWithApi<B, BE>,
-        P: WebbGadgetProtocol<B, BE, C>,
-    > ProtocolConfig for EcdsaProtocolConfig<N, P, C, B, BE>
-where
-    <C as ProvideRuntimeApi<B>>::Api: JobsApi<B, AccountId>,
-{
-    type Network = N;
-    type Block = B;
-    type Backend = BE;
-    type Protocol = P;
-    type Client = C;
-
-    fn take_network(&mut self) -> Self::Network {
-        self.network.take().expect("Network not set")
-    }
-
-    fn take_protocol(&mut self) -> Self::Protocol {
-        self.protocol.take().expect("Protocol not set")
-    }
-
-    fn take_client(&mut self) -> Self::Client {
-        self.client.take().expect("Client not set")
-    }
-}
+define_protocol!(EcdsaProtocolConfig);
